Added documentation with minor refactoring
