+++ lib/net/NetworkMonitor.ts
