+++ lib/ui/ExpandableMenuItem.ts
