+++ lib/utils/ExtensionUtils.ts
